+++ all/be_memo_pair/index.d.ts
@@ -4,13 +4,13 @@ export declare function be_memo_pair_<
 	val_T,
 	ns_T extends string = '',
 	_memo_T extends memo_T<val_T> = memo_T<val_T>,
-	ctx_T extends Ctx = Ctx_wide_T<ns_T>,
+	ctx_T extends Ctx_wide_T<ns_T> = Ctx_wide_T<ns_T>,
 >(be:Be<_memo_T, ns_T, ctx_T>):be_memo_pair_T<val_T, ns_T, _memo_T, ctx_T>
 export declare function be_memo_pair_<
 	val_T,
 	ns_T extends string = '',
 	_memo_T extends memo_T<val_T> = memo_T<val_T>,
-	ctx_T extends Ctx = Ctx_wide_T<ns_T>
+	ctx_T extends Ctx_wide_T<ns_T> = Ctx_wide_T<ns_T>,
 >(
 	rmemo__new:(ctx:ctx_T, memo:_memo_T&{ _:val_T })=>val_T,
 	config?:be_config_T
@@ -19,10 +19,10 @@ export type be_memo_pair_T<
 	val_T,
 	ns_T extends string = '',
 	_memo_T extends memo_T<val_T> = memo_T<val_T>,
-	ctx_T extends Ctx = Ctx_wide_T<ns_T>
+	ctx_T extends Ctx_wide_T<ns_T> = Ctx_wide_T<ns_T>,
 > = [
-	ctx__be_T<ctx_T, _memo_T, ns_T>,
-	ctx__get_T<ctx_T, val_T>,
+	ctx__be_T<_memo_T, ns_T, ctx_T>,
+	ctx__get_T<val_T, ns_T, ctx_T>,
 ]&{
 	add<add_val_T>(
 		add_def:(ctx:ctx_T, sig:sig_T<val_T>, prev_val:add_val_T|undefined)=>add_val_T