+++ all/be_memosig_triple/index.d.ts
@@ -4,13 +4,13 @@ export declare function be_memosig_triple_<
 	val_T,
 	ns_T extends string = '',
 	_sig_T extends sig_T<val_T> = sig_T<val_T>,
-	ctx_T extends Ctx = Ctx_wide_T<ns_T>,
+	ctx_T extends Ctx_wide_T<ns_T> = Ctx_wide_T<ns_T>,
 >(be:Be<_sig_T, ns_T, ctx_T>):be_memosig_triple_T<val_T, ns_T, _sig_T, ctx_T>
 export declare function be_memosig_triple_<
 	val_T,
 	ns_T extends string = '',
 	_sig_T extends sig_T<val_T> = sig_T<val_T>,
-	ctx_T extends Ctx = Ctx_wide_T<ns_T>
+	ctx_T extends Ctx_wide_T<ns_T> = Ctx_wide_T<ns_T>,
 >(
 	rmemo__new:(ctx:ctx_T, memosig:_sig_T)=>val_T,
 	config?:be_config_T<ns_T>
@@ -19,11 +19,11 @@ export type be_memosig_triple_T<
 	val_T,
 	ns_T extends string = '',
 	_sig_T extends sig_T<val_T> = sig_T<val_T>,
-	ctx_T extends Ctx = Ctx_wide_T<ns_T>
+	ctx_T extends Ctx_wide_T<ns_T> = Ctx_wide_T<ns_T>,
 > = [
-	ctx__be_T<ctx_T, _sig_T, ns_T>,
-	ctx__get_T<ctx_T, val_T>,
-	ctx__set_T<ctx_T, val_T>,
+	ctx__be_T<_sig_T, ns_T, ctx_T>,
+	ctx__get_T<val_T, ns_T, ctx_T>,
+	ctx__set_T<val_T, ns_T, ctx_T>,
 ]&{
 	add<add_val_T>(
 		add_def:(ctx:ctx_T, sig:_sig_T, prev_val:add_val_T|undefined)=>add_val_T